+++ src/SFA.DAS.Roatp.Application/Common/LatLongValidator.cs
@@ -13,19 +13,19 @@ public class LatLongValidator : AbstractValidator<ILatLon>
     {
         RuleFor(p => p.Latitude.HasValue)
             .Equal(false)
-            .When(p => p.Latitude > (double)Domain.Constants.NationalLatLong.MaximumLatitude)
+            .When(p => p.Latitude > Domain.Constants.NationalLatLong.MaximumLatitude)
             .WithMessage(LatitudeOutsideUk);
         RuleFor(p => p.Latitude.HasValue)
             .Equal(false)
-            .When(p => p.Latitude < (double)Domain.Constants.NationalLatLong.MinimumLatitude)
+            .When(p => p.Latitude < Domain.Constants.NationalLatLong.MinimumLatitude)
             .WithMessage(LatitudeOutsideUk);
         RuleFor(p => p.Longitude.HasValue)
             .Equal(false)
-            .When(p => p.Longitude > (double)Domain.Constants.NationalLatLong.MaximumLongitude)
+            .When(p => p.Longitude > Domain.Constants.NationalLatLong.MaximumLongitude)
             .WithMessage(LongitudeOutsideUk);
         RuleFor(p => p.Longitude.HasValue)
             .Equal(false)
-            .When(p => p.Longitude < (double)Domain.Constants.NationalLatLong.MinimumLongitude)
+            .When(p => p.Longitude < Domain.Constants.NationalLatLong.MinimumLongitude)
             .WithMessage(LongitudeOutsideUk);
         RuleFor(p => p.Longitude.HasValue && !p.Latitude.HasValue)
             .Equal(false)
